+++ src/micro_app_element.ts
@@ -6,6 +6,9 @@ import type {
   OptionsType,
   NormalKey,
 } from '@micro-app/types'
+import microApp from './micro_app'
+import dispatchLifecyclesEvent from './interact/lifecycles_event'
+import globalEnv from './libs/global_env'
 import {
   defer,
   formatAppName,
@@ -23,19 +26,22 @@ import {
   ObservedAttrName,
   lifeCycles,
 } from './constants'
-import CreateApp, { appInstanceMap } from './create_app'
-import microApp from './micro_app'
-import dispatchLifecyclesEvent from './interact/lifecycles_event'
-import globalEnv from './libs/global_env'
-import { getNoHashMicroPathFromURL, router } from './sandbox/router'
+import CreateApp, {
+  appInstanceMap,
+} from './create_app'
+import {
+  router,
+  getNoHashMicroPathFromURL,
+  getRouterMode,
+} from './sandbox/router'
 
 /**
  * define element
  * @param tagName element name
  */
-export function defineElement(tagName: string): void {
+export function defineElement (tagName: string): void {
   class MicroAppElement extends HTMLElement implements MicroAppElementType {
-    static get observedAttributes(): string[] {
+    static get observedAttributes (): string[] {
       return ['name', 'url']
     }
 
@@ -59,7 +65,7 @@ export function defineElement(tagName: string): void {
     // baseRoute: route prefix, default is ''
     // keep-alive: open keep-alive mode
 
-    public connectedCallback(): void {
+    public connectedCallback (): void {
       const cacheCount = ++this.connectedCount
       this.connectStateMap.set(cacheCount, true)
       /**
@@ -84,7 +90,7 @@ export function defineElement(tagName: string): void {
       })
     }
 
-    public disconnectedCallback(): void {
+    public disconnectedCallback (): void {
       this.connectStateMap.set(this.connectedCount, false)
       this.handleDisconnected()
     }
@@ -93,7 +99,7 @@ export function defineElement(tagName: string): void {
      * Re render app from the command line
      * MicroAppElement.reload(destroy)
      */
-    public reload(destroy?: boolean): Promise<boolean> {
+    public reload (destroy?: boolean): Promise<boolean> {
       return new Promise((resolve) => {
         const handleAfterReload = () => {
           this.removeEventListener(lifeCycles.MOUNTED, handleAfterReload)
@@ -117,7 +123,7 @@ export function defineElement(tagName: string): void {
      * common action for unmount
      * @param destroy reload param
      */
-    private handleDisconnected(destroy = false, callback?: CallableFunction): void {
+    private handleDisconnected (destroy = false, callback?: CallableFunction): void {
       const app = appInstanceMap.get(this.appName)
       if (app && !app.isUnmounted() && !app.isHidden()) {
         // keep-alive
@@ -129,7 +135,7 @@ export function defineElement(tagName: string): void {
       }
     }
 
-    public attributeChangedCallback(attr: ObservedAttrName, _oldVal: string, newVal: string): void {
+    public attributeChangedCallback (attr: ObservedAttrName, _oldVal: string, newVal: string): void {
       if (
         this.legalAttribute(attr, newVal) &&
         this[attr === ObservedAttrName.NAME ? 'appName' : 'appUrl'] !== newVal
@@ -166,14 +172,14 @@ export function defineElement(tagName: string): void {
     }
 
     // handle for connectedCallback run before attributeChangedCallback
-    private handleInitialNameAndUrl(): void {
+    private handleInitialNameAndUrl (): void {
       this.connectStateMap.get(this.connectedCount) && this.handleConnected()
     }
 
     /**
      * first mount of this app
      */
-    private handleConnected(): void {
+    private handleConnected (): void {
       if (!this.appName || !this.appUrl) return
 
       if (this.getDisposeResult('shadowDOM') && !this.shadowRoot && isFunction(this.attachShadow)) {
@@ -263,7 +269,7 @@ export function defineElement(tagName: string): void {
     }
 
     // remount app or create app if attribute url or name change
-    private actionsForAttributeChange(
+    private actionsForAttributeChange (
       formatAttrName: string,
       formatAttrUrl: string,
       oldApp: AppInterface | void,
@@ -275,7 +281,7 @@ export function defineElement(tagName: string): void {
 
       this.appName = formatAttrName
       this.appUrl = formatAttrUrl
-        ; (this.shadowRoot ?? this).innerHTML = ''
+      ; (this.shadowRoot ?? this).innerHTML = ''
       if (formatAttrName !== this.getAttribute('name')) {
         this.setAttribute('name', this.appName)
       }
@@ -324,7 +330,7 @@ export function defineElement(tagName: string): void {
      * @param name attribute name
      * @param val attribute value
      */
-    private legalAttribute(name: string, val: AttrType): boolean {
+    private legalAttribute (name: string, val: AttrType): boolean {
       if (!isString(val) || !val) {
         logError(`unexpected attribute ${name}, please check again`, this.appName)
 
@@ -335,15 +341,15 @@ export function defineElement(tagName: string): void {
     }
 
     // create app instance
-    private handleCreateApp(): void {
+    private handleCreateApp (): void {
       const createAppInstance = () => new CreateApp({
         name: this.appName,
         url: this.appUrl,
+        container: this.shadowRoot ?? this,
         scopecss: this.useScopecss(),
         useSandbox: this.useSandbox(),
         inline: this.getDisposeResult('inline'),
         iframe: this.getDisposeResult('iframe'),
-        container: this.shadowRoot ?? this,
         ssrUrl: this.ssrUrl,
       })
 
@@ -374,7 +380,7 @@ export function defineElement(tagName: string): void {
      * 2. is remount in another container ?
      * 3. is remount with change properties of the container ?
      */
-    private handleMount(app: AppInterface): void {
+    private handleMount (app: AppInterface): void {
       app.isPrefetch = false
       // TODO: Can defer be removed?
       defer(() => this.mount(app))
@@ -383,16 +389,15 @@ export function defineElement(tagName: string): void {
     /**
      * public mount action for micro_app_element & create_app
      */
-    public mount(app: AppInterface): void {
+    public mount (app: AppInterface): void {
       app.mount({
         container: this.shadowRoot ?? this,
         inline: this.getDisposeResult('inline'),
-        useMemoryRouter: !this.getDisposeResult('disable-memory-router'),
-        defaultPage: this.getDefaultPageValue(),
+        routerMode: this.getMemoryRouterMode(),
         baseroute: this.getBaseRouteCompatible(),
+        defaultPage: this.getDefaultPage(),
         disablePatchRequest: this.getDisposeResult('disable-patch-request'),
         fiber: this.getDisposeResult('fiber'),
-        // hiddenRouter: this.getDisposeResult('hidden-router'),
       })
     }
 
@@ -401,7 +406,7 @@ export function defineElement(tagName: string): void {
      * @param destroy delete cache resources when unmount
      * @param unmountcb callback
      */
-    public unmount(destroy?: boolean, unmountcb?: CallableFunction): void {
+    public unmount (destroy?: boolean, unmountcb?: CallableFunction): void {
       const app = appInstanceMap.get(this.appName)
       if (app && !app.isUnmounted()) {
         app.unmount({
@@ -414,7 +419,7 @@ export function defineElement(tagName: string): void {
     }
 
     // hidden app when disconnectedCallback called with keep-alive
-    private handleHiddenKeepAliveApp(callback?: CallableFunction): void {
+    private handleHiddenKeepAliveApp (callback?: CallableFunction): void {
       const app = appInstanceMap.get(this.appName)
       if (app && !app.isUnmounted() && !app.isHidden()) {
         app.hiddenKeepAliveApp(callback)
@@ -422,7 +427,7 @@ export function defineElement(tagName: string): void {
     }
 
     // show app when connectedCallback called with keep-alive
-    private handleShowKeepAliveApp(app: AppInterface): void {
+    private handleShowKeepAliveApp (app: AppInterface): void {
       // must be async
       defer(() => app.showKeepAliveApp(this.shadowRoot ?? this))
     }
@@ -432,12 +437,12 @@ export function defineElement(tagName: string): void {
      * Global setting is lowest priority
      * @param name Configuration item name
      */
-    private getDisposeResult<T extends keyof OptionsType>(name: T): boolean {
+    public getDisposeResult <T extends keyof OptionsType> (name: T): boolean {
       return (this.compatibleProperties(name) || !!microApp.options[name]) && this.compatibleDisableProperties(name)
     }
 
     // compatible of disableScopecss & disableSandbox
-    private compatibleProperties(name: string): boolean {
+    private compatibleProperties (name: string): boolean {
       if (name === 'disable-scopecss') {
         return this.hasAttribute('disable-scopecss') || this.hasAttribute('disableScopecss')
       } else if (name === 'disable-sandbox') {
@@ -447,7 +452,7 @@ export function defineElement(tagName: string): void {
     }
 
     // compatible of disableScopecss & disableSandbox
-    private compatibleDisableProperties(name: string): boolean {
+    private compatibleDisableProperties (name: string): boolean {
       if (name === 'disable-scopecss') {
         return this.getAttribute('disable-scopecss') !== 'false' && this.getAttribute('disableScopecss') !== 'false'
       } else if (name === 'disable-sandbox') {
@@ -456,18 +461,18 @@ export function defineElement(tagName: string): void {
       return this.getAttribute(name) !== 'false'
     }
 
-    private useScopecss(): boolean {
+    private useScopecss (): boolean {
       return !(this.getDisposeResult('disable-scopecss') || this.getDisposeResult('shadowDOM'))
     }
 
-    private useSandbox(): boolean {
+    private useSandbox (): boolean {
       return !this.getDisposeResult('disable-sandbox')
     }
 
     /**
      * Determine whether the core options of the existApp is consistent with the new one
      */
-    private sameCoreOptions(app: AppInterface): boolean {
+    private sameCoreOptions (app: AppInterface): boolean {
       return (
         app.scopecss === this.useScopecss() &&
         app.useSandbox === this.useSandbox() &&
@@ -480,34 +485,35 @@ export function defineElement(tagName: string): void {
      * get baseRoute
      * getAttribute('baseurl') is compatible writing of versions below 0.3.1
      */
-    private getBaseRouteCompatible(): string {
+    private getBaseRouteCompatible (): string {
       return this.getAttribute('baseroute') ?? this.getAttribute('baseurl') ?? ''
     }
 
     // compatible of destroy
-    private getDestroyCompatibleResult(): boolean {
+    private getDestroyCompatibleResult (): boolean {
       return this.getDisposeResult('destroy') || this.getDisposeResult('destory')
     }
 
     /**
      * destroy has priority over destroy keep-alive
      */
-    private getKeepAliveModeResult(): boolean {
+    private getKeepAliveModeResult (): boolean {
       return this.getDisposeResult('keep-alive') && !this.getDestroyCompatibleResult()
     }
 
     /**
      * change ssrUrl in ssr mode
      */
-    private updateSsrUrl(baseUrl: string): void {
+    private updateSsrUrl (baseUrl: string): void {
       if (this.getDisposeResult('ssr')) {
+        // TODO: disable-memory-router不存在了，这里需要更新一下
         if (this.getDisposeResult('disable-memory-router') || this.getDisposeResult('disableSandbox')) {
           const rawLocation = globalEnv.rawWindow.location
           this.ssrUrl = CompletionPath(rawLocation.pathname + rawLocation.search, baseUrl)
         } else {
           // get path from browser URL
           let targetPath = getNoHashMicroPathFromURL(this.appName, baseUrl)
-          const defaultPagePath = this.getDefaultPageValue()
+          const defaultPagePath = this.getDefaultPage()
           if (!targetPath && defaultPagePath) {
             const targetLocation = createURL(defaultPagePath, baseUrl)
             targetPath = targetLocation.origin + targetLocation.pathname + targetLocation.search
@@ -522,7 +528,7 @@ export function defineElement(tagName: string): void {
     /**
      * get config of default page
      */
-    private getDefaultPageValue(): string {
+    private getDefaultPage (): string {
       return (
         router.getDefaultPage(this.appName) ||
         this.getAttribute('default-page') ||
@@ -532,11 +538,19 @@ export function defineElement(tagName: string): void {
     }
 
     /**
-     * Rewrite micro-app.setAttribute, process attr data
+     * get config of router-mode
+     * @returns router-mode
+     */
+    private getMemoryRouterMode () : string {
+      return getRouterMode(this.getAttribute('router-mode'), this)
+    }
+
+    /**
+     * rewrite micro-app.setAttribute, process attr data
      * @param key attr name
      * @param value attr value
      */
-    public setAttribute(key: string, value: any): void {
+    public setAttribute (key: string, value: any): void {
       if (key === 'data') {
         if (isPlainObject(value)) {
           const cloneValue: Record<NormalKey, unknown> = {}
@@ -557,7 +571,7 @@ export function defineElement(tagName: string): void {
     /**
      * Data from the base application
      */
-    set data(value: Record<PropertyKey, unknown> | null) {
+    set data (value: Record<PropertyKey, unknown> | null) {
       if (this.appName) {
         microApp.setData(this.appName, value as Record<PropertyKey, unknown>)
       } else {
@@ -568,7 +582,7 @@ export function defineElement(tagName: string): void {
     /**
      * get data only used in jsx-custom-event once
      */
-    get data(): Record<PropertyKey, unknown> | null {
+    get data (): Record<PropertyKey, unknown> | null {
       if (this.appName) {
         return microApp.getData(this.appName, true)
       } else if (this.cacheData) {